fix(tests): move xorq perf smoke tests behind the importorskip guard (3.14 has no xorq)
